Bug 13261 - Better check in message for patrons with indefinite restricition
This patch adds a better check in message for patrons with indefinite restriction.

To test:
Check out an item to a patron.
Add a manual restriction without expiry date to that patron.
Check in the item.

Without patch, the checkin message reads:
Reminder: Patron was earlier restricted until 9999-12-31

Apply patch and repeat steps above.
The message should now read:
Reminder: Patron has a restriction (no expiry date)

NOTE: Changed wording at two places following Owen's  suggestion.  New: "Patron
has an indefinite restriction"


Thanks Marc for catching this case. I was thinking like you that the wording
sounded strange while playing with bug 13242. Merge the original patch and the
followup, containing a better wording, thanks to Owen comment.

Works as described, no problems found.
Passes tests and QA script.

+++ C4/Circulation.pm
@@ -1946,11 +1946,15 @@ sub AddReturn {
             }
         # there's no overdue on the item but borrower had been previously debarred
         } elsif ( $issue->{date_due} and $borrower->{'debarred'} ) {
-             my $borrower_debar_dt = dt_from_string( $borrower->{debarred} );
-             $borrower_debar_dt->truncate(to => 'day');
-             my $today_dt = $today->clone()->truncate(to => 'day');
-             if ( DateTime->compare( $borrower_debar_dt, $today_dt ) != -1 ) {
-                 $messages->{'PrevDebarred'} = $borrower->{'debarred'};
+             if ( $borrower->{debarred} eq "9999-12-31") {
+                $messages->{'ForeverDebarred'} = $borrower->{'debarred'};
+             } else {
+                  my $borrower_debar_dt = dt_from_string( $borrower->{debarred} );
+                  $borrower_debar_dt->truncate(to => 'day');
+                  my $today_dt = $today->clone()->truncate(to => 'day');
+                  if ( DateTime->compare( $borrower_debar_dt, $today_dt ) != -1 ) {
+                      $messages->{'PrevDebarred'} = $borrower->{'debarred'};
+                  }
              }
         }
     }
